# api/http/openapi.yaml
openapi: 3.1.0
info:
  title: MatchLab Gateway API
  version: 2.3.0
  description: |
    Текущий публичный HTTP API MatchLab: учётные записи, сеансы, подтверждение почты и файлы.
    Это описание реализованных маршрутов Gateway. Для локальной проверки сначала
    зарегистрируйтесь, подтвердите email из письма, затем войдите и укажите полученный токен доступа в Bearer Auth.
    Для проверки почты сначала добавьте разрешённые домены HSE локальным seed-скриптом.
    Письмо с подтверждением появится в Mailpit по адресу http://localhost:8025.
    Токен из ссылки в письме передайте в метод подтверждения адреса.
servers:
  - url: http://localhost:8080
    description: Gateway в локальном Docker Compose
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Учётная запись]
      summary: Зарегистрировать пользователя
      description: Email обязателен и приводится к нижнему регистру. Сохраняется заявка и отправляется письмо; аккаунта и сеанса ещё нет. Повтор с тем же email заменяет пароль и токен заявки. Существующий аккаунт менять нельзя.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
            example:
              email: student@hse.ru
              password: example-password-123
      responses:
        '201':
          description: Письмо принято SMTP. Для продолжения подтвердите email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResult'
              example:
                status: pending
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email уже зарегистрирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: email уже зарегистрирован
        '403':
          $ref: '#/components/responses/OrganizationUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/auth/login:
    post:
      tags: [Учётная запись]
      summary: Войти по email и паролю
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
            example:
              email: student@hse.ru
              password: example-password-123
      responses:
        '200':
          description: Сеанс создан. Используйте access_token для защищённых методов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
              example:
                user_id: 4f9a4c95-6144-4ec8-89e8-3866207d7561
                access_token: пример-токена-доступа
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Неверный email или пароль, в том числе до подтверждения регистрации.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: неверный email или пароль
        '403':
          description: Почта не подтверждена или учётная запись неактивна.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: подтвердите адрес электронной почты
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/users/me/email:
    get:
      tags: [Почта]
      summary: Получить состояние адреса
      operationId: getEmailStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Текущий и ожидающий адреса пользователя.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailStatus'
              example:
                status: verified
                email: student@hse.ru
        '403':
          description: Почта не подтверждена (включая старые сеансы).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/auth/email/confirm:
    post:
      tags: [Почта]
      summary: Подтвердить адрес из письма
      description: |
        Передайте token из актуальной ссылки в письме. Этот метод не требует Bearer-токена.
        Аккаунт создаётся с паролем подтверждённой заявки, а заявка удаляется.
        Старые ссылки после повторной регистрации и использованные токены недействительны.
      operationId: confirmEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmEmailRequest'
            example:
              token: ТОКЕН_ИЗ_ПИСЬМА
      responses:
        '204':
          description: Адрес подтверждён, аккаунт создан.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/OrganizationUnavailable'
        '404':
          description: Токен не найден или уже использован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: токен подтверждения не найден или уже использован
        '409':
          description: Email уже зарегистрирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: email уже зарегистрирован
        '410':
          description: Срок действия токена истёк.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: срок подтверждения адреса электронной почты истёк
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/sessions/current:
    get:
      tags: [Сеанс]
      summary: Получить пользователя текущего сеанса
      operationId: getCurrentSession
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Сеанс действителен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserId'
              example:
                user_id: 4f9a4c95-6144-4ec8-89e8-3866207d7561
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Почта не подтверждена.
        '503':
          $ref: '#/components/responses/Unavailable'
    delete:
      tags: [Сеанс]
      summary: Завершить текущий сеанс
      operationId: revokeCurrentSession
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Сеанс отозван.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/Unavailable'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Токен access_token из ответа POST /api/v1/auth/login.
  schemas:
...
    Credentials:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: Обязательный email регистрации и входа, нормализуется в нижний регистр.
          maxLength: 320
        password:
          type: string
          format: password
          description: От 8 до 72 байт в UTF-8.
    UserId:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
    RegistrationResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending]
    LoginResult:
      type: object
      required: [user_id, access_token]
      properties:
        user_id:
          type: string
          format: uuid
        access_token:
          type: string
    ConfirmEmailRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          description: Одноразовый токен из письма, действующий 30 минут.
    EmailStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, verified]
        email:
          type: string
          format: email
          description: Текущий подтверждённый адрес, если есть.
        pending_email:
          type: string
          format: email
          description: Email регистрации, ожидающий подтверждения.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
...
    BadRequest:
      description: Некорректный JSON или значения запроса.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: некорректное тело запроса
    Unauthorized:
      description: Токен доступа отсутствует или недействителен.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: недействительный сеанс
    OrganizationUnavailable:
      description: Почтовый домен не принадлежит активной разрешённой организации.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: подтверждение почты для этой организации недоступно
    Unavailable:
      description: Зависимый сервис или SMTP-сервер недоступен.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: сервис недоступен